update to TagFavourites
